fix: Rename slack.js to index.js to follow Nodejs convention. (#12)
